update on connecting API
