--- src/pages/NoPage.jsx
@@ -1,10 +1,0 @@
-import React from "react";
-const NoPage = () => {
-  return (
-    <>
-      <h1>404</h1>
-    </>
-  );
-};
-
-export default NoPage;
